factored out print function into helper file, and extended it to support int and double printing
