+++ CMakeLists.txt
@@ -9,7 +9,9 @@ add_executable(dct
                ${PROJECT_SOURCE_DIR}/src/dct.c
                ${PROJECT_SOURCE_DIR}/src/dct/naive.c
                ${PROJECT_SOURCE_DIR}/src/dct/twostep_slow.c
-               ${PROJECT_SOURCE_DIR}/src/dct/loeffler_float.c)
+               ${PROJECT_SOURCE_DIR}/src/dct/loeffler_float.c
+               ${PROJECT_SOURCE_DIR}/src/util/helpers.c
+               )
 target_include_directories(dct PUBLIC ${PROJECT_SOURCE_DIR}/include)
